+++ bigframes/dataframe.py
@@ -19,6 +19,7 @@ from __future__ import annotations
 import datetime
 import inspect
 import itertools
+import json
 import re
 import sys
 import textwrap
@@ -179,9 +180,7 @@ class DataFrame(vendored_pandas_frame.DataFrame):
             if columns:
                 block = block.select_columns(list(columns))  # type:ignore
             if dtype:
-                block = block.multi_apply_unary_op(
-                    block.value_columns, ops.AsTypeOp(to_type=dtype)
-                )
+                block = block.multi_apply_unary_op(ops.AsTypeOp(to_type=dtype))
             self._block = block
 
         else:
@@ -743,16 +742,14 @@ class DataFrame(vendored_pandas_frame.DataFrame):
 
         df = self.copy()
         if bigframes.options.experiments.blob:
-            import bigframes.bigquery as bbq
-
             blob_cols = [
                 col
                 for col in df.columns
                 if df[col].dtype == bigframes.dtypes.OBJ_REF_DTYPE
             ]
             for col in blob_cols:
-                df[col] = df[col]._apply_unary_op(ops.ObjGetAccessUrl(mode="R"))
-                df[col] = bbq.json_extract(df[col], "$.access_urls.read_url")
+                # TODO(garrettwu): Not necessary to get access urls for all the rows. Update when having a to get URLs from local data.
+                df[col] = df[col].blob._get_runtime(mode="R", with_metadata=True)
 
         # TODO(swast): pass max_columns and get the true column count back. Maybe
         # get 1 more column than we have requested so that pandas can add the
@@ -769,11 +766,21 @@ class DataFrame(vendored_pandas_frame.DataFrame):
             # Allows to preview images in the DataFrame. The implementation changes the string repr as well, that it doesn't truncate strings or escape html charaters such as "<" and ">". We may need to implement a full-fledged repr module to better support types not in pandas.
             if bigframes.options.experiments.blob:
 
-                def url_to_image_html(url: str) -> str:
-                    # url is a json string, which already contains double-quotes ""
-                    return f"<img src={url}>"
+                def obj_ref_rt_to_html(obj_ref_rt) -> str:
+                    obj_ref_rt_json = json.loads(obj_ref_rt)
+                    content_type = typing.cast(
+                        str,
+                        obj_ref_rt_json["objectref"]["details"]["gcs_metadata"][
+                            "content_type"
+                        ],
+                    )
+                    if content_type.startswith("image"):
+                        url = obj_ref_rt_json["access_urls"]["read_url"]
+                        return f'<img src="{url}">'
 
-                formatters = {blob_col: url_to_image_html for blob_col in blob_cols}
+                    return f'uri: {obj_ref_rt_json["objectref"]["uri"]}, authorizer: {obj_ref_rt_json["objectref"]["authorizer"]}'
+
+                formatters = {blob_col: obj_ref_rt_to_html for blob_col in blob_cols}
 
                 # set max_colwidth so not to truncate the image url
                 with pandas.option_context("display.max_colwidth", None):
@@ -846,9 +853,7 @@ class DataFrame(vendored_pandas_frame.DataFrame):
                 left_input=ex.free_var("var1"),
                 right_input=ex.const(other),
             )
-        return DataFrame(
-            self._block.multi_apply_unary_op(self._block.value_columns, expr)
-        )
+        return DataFrame(self._block.multi_apply_unary_op(expr))
 
     def _apply_series_binop_axis_0(
         self,
@@ -1467,6 +1472,48 @@ class DataFrame(vendored_pandas_frame.DataFrame):
         result.columns = orig_columns
 
         return result
+
+    def corrwith(
+        self,
+        other: typing.Union[DataFrame, bigframes.series.Series],
+        *,
+        numeric_only: bool = False,
+    ):
+        other_frame = other if isinstance(other, DataFrame) else other.to_frame()
+        if numeric_only:
+            l_frame = self._drop_non_numeric()
+            r_frame = other_frame._drop_non_numeric()
+        else:
+            l_frame = self._raise_on_non_numeric("corrwith")
+            r_frame = other_frame._raise_on_non_numeric("corrwith")
+
+        l_block = l_frame.astype(bigframes.dtypes.FLOAT_DTYPE)._block
+        r_block = r_frame.astype(bigframes.dtypes.FLOAT_DTYPE)._block
+
+        if isinstance(other, DataFrame):
+            block, labels, expr_pairs = l_block._align_both_axes(r_block, how="inner")
+        else:
+            assert isinstance(other, bigframes.series.Series)
+            block, labels, expr_pairs = l_block._align_axis_0(r_block, how="inner")
+
+        na_cols = l_block.column_labels.join(
+            r_block.column_labels, how="outer"
+        ).difference(labels)
+
+        block, _ = block.aggregate(
+            aggregations=tuple(
+                ex.BinaryAggregation(agg_ops.CorrOp(), left_ex, right_ex)
+                for left_ex, right_ex in expr_pairs
+            ),
+            column_labels=labels,
+        )
+        block = block.project_exprs(
+            (ex.const(float("nan")),) * len(na_cols), labels=na_cols
+        )
+        block = block.transpose(
+            original_row_index=pandas.Index([None]), single_row_mode=True
+        )
+        return bigframes.pandas.Series(block)
 
     def to_arrow(
         self,
@@ -2360,6 +2407,9 @@ class DataFrame(vendored_pandas_frame.DataFrame):
         result.columns.names = self.columns.names
         return result
 
+    def mask(self, cond, other=None):
+        return self.where(~cond, other=other)
+
     def dropna(
         self,
         *,
@@ -2401,9 +2451,7 @@ class DataFrame(vendored_pandas_frame.DataFrame):
                 result = result.reset_index()
             return DataFrame(result)
         else:
-            isnull_block = self._block.multi_apply_unary_op(
-                self._block.value_columns, ops.isnull_op
-            )
+            isnull_block = self._block.multi_apply_unary_op(ops.isnull_op)
             if how == "any":
                 null_locations = DataFrame(isnull_block).any().to_pandas()
             else:  # 'all'
@@ -3040,8 +3088,15 @@ class DataFrame(vendored_pandas_frame.DataFrame):
         return DataFrame(block)
 
     def join(
-        self, other: DataFrame, *, on: Optional[str] = None, how: str = "left"
+        self,
+        other: Union[DataFrame, bigframes.series.Series],
+        *,
+        on: Optional[str] = None,
+        how: str = "left",
     ) -> DataFrame:
+        if isinstance(other, bigframes.series.Series):
+            other = other.to_frame()
+
         left, right = self, other
 
         if not left.columns.intersection(right.columns).empty:
@@ -3829,7 +3884,7 @@ class DataFrame(vendored_pandas_frame.DataFrame):
         return as_pandas_default_index.to_orc(path, **kwargs)
 
     def _apply_unary_op(self, operation: ops.UnaryOp) -> DataFrame:
-        block = self._block.multi_apply_unary_op(self._block.value_columns, operation)
+        block = self._block.multi_apply_unary_op(operation)
         return DataFrame(block)
 
     def _map_clustering_columns(
@@ -3924,6 +3979,10 @@ class DataFrame(vendored_pandas_frame.DataFrame):
         )
 
     def apply(self, func, *, axis=0, args: typing.Tuple = (), **kwargs):
+        # In Bigframes remote function, DataFrame '.apply' method is specifically
+        # designed to work with row-wise or column-wise operations, where the input
+        # to the applied function should be a Series, not a scalar.
+
         if utils.get_axis_number(axis) == 1:
             msg = "axis=1 scenario is in preview."
             warnings.warn(msg, category=bfe.PreviewWarning)
@@ -3937,7 +3996,7 @@ class DataFrame(vendored_pandas_frame.DataFrame):
                 # Early check whether the dataframe dtypes are currently supported
                 # in the remote function
                 # NOTE: Keep in sync with the value converters used in the gcf code
-                # generated in remote_function_template.py
+                # generated in function_template.py
                 remote_function_supported_dtypes = (
                     bigframes.dtypes.INT_DTYPE,
                     bigframes.dtypes.FLOAT_DTYPE,
@@ -4031,8 +4090,19 @@ class DataFrame(vendored_pandas_frame.DataFrame):
 
             return result_series
 
+        # At this point column-wise or element-wise remote function operation will
+        # be performed (not supported).
+        if hasattr(func, "bigframes_remote_function"):
+            raise NotImplementedError(
+                "BigFrames DataFrame '.apply()' does not support remote function "
+                "for column-wise (i.e. with axis=0) operations, please use a "
+                "regular python function instead. For element-wise operations of "
+                "the remote function, please use '.map()'."
+            )
+
         # Per-column apply
         results = {name: func(col, *args, **kwargs) for name, col in self.items()}
+
         if all(
             [
                 isinstance(val, bigframes.series.Series) or utils.is_list_like(val)
@@ -4242,6 +4312,56 @@ class DataFrame(vendored_pandas_frame.DataFrame):
     @property
     def plot(self):
         return plotting.PlotAccessor(self)
+
+    def hist(
+        self, by: typing.Optional[typing.Sequence[str]] = None, bins: int = 10, **kwargs
+    ):
+        return self.plot.hist(by=by, bins=bins, **kwargs)
+
+    hist.__doc__ = inspect.getdoc(plotting.PlotAccessor.hist)
+
+    def line(
+        self,
+        x: typing.Optional[typing.Hashable] = None,
+        y: typing.Optional[typing.Hashable] = None,
+        **kwargs,
+    ):
+        return self.plot.line(x=x, y=y, **kwargs)
+
+    line.__doc__ = inspect.getdoc(plotting.PlotAccessor.line)
+
+    def area(
+        self,
+        x: typing.Optional[typing.Hashable] = None,
+        y: typing.Optional[typing.Hashable] = None,
+        stacked: bool = True,
+        **kwargs,
+    ):
+        return self.plot.area(x=x, y=y, stacked=stacked, **kwargs)
+
+    area.__doc__ = inspect.getdoc(plotting.PlotAccessor.area)
+
+    def bar(
+        self,
+        x: typing.Optional[typing.Hashable] = None,
+        y: typing.Optional[typing.Hashable] = None,
+        **kwargs,
+    ):
+        return self.plot.bar(x=x, y=y, **kwargs)
+
+    bar.__doc__ = inspect.getdoc(plotting.PlotAccessor.bar)
+
+    def scatter(
+        self,
+        x: typing.Optional[typing.Hashable] = None,
+        y: typing.Optional[typing.Hashable] = None,
+        s: typing.Union[typing.Hashable, typing.Sequence[typing.Hashable]] = None,
+        c: typing.Union[typing.Hashable, typing.Sequence[typing.Hashable]] = None,
+        **kwargs,
+    ):
+        return self.plot.scatter(x=x, y=y, s=s, c=c, **kwargs)
+
+    scatter.__doc__ = inspect.getdoc(plotting.PlotAccessor.scatter)
 
     def __matmul__(self, other) -> DataFrame:
         return self.dot(other)